#660 - store - split cart, attribute modules
