+++ src/store/modules/attribute/getters.js
@@ -1,0 +1,4 @@
+export default {
+  attributeListByCode: (state) => state.list_by_code,
+  attributeListById: (state) => state.list_by_id
+}
