+++ src/store/modules/attribute/index.js
@@ -1,0 +1,15 @@
+import actions from './actions'
+import getters from './getters'
+import mutations from './mutations'
+
+export default {
+  namespaced: true,
+  state: {
+    list_by_code: {},
+    list_by_id: {},
+    labels: {}
+  },
+  getters,
+  actions,
+  mutations
+}
